Use indirect loads and stores only for aggregates
Previously, we used a single-element array for local variables,
requiring indirect loads and stores for locals. This made it
impossible to see through local variable accesses and the
resulting code had significant overhead when accessing locals.

We now have a mechanism to define locals and generate symbol
names for them. To keep track of the names to the allocations,
we use a mapping from the AllocaInst to the variable name.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>